Properly generate secrets files
This allows for the new secrets variables to be added to the
existing secrets files without leaving any of the new secrets
ungenerated. This is necessary because we introduce new secrets
in Mitaka. The old method would just leave out the new ones
causing errors.

Connects http://github.com/rcbops/u-suk-dev#475